requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
fastapi==0.79.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
passlib==1.7.4
protobuf==4.21.4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
watchfiles==0.16.1
websockets==10.3
